refactor: live 방송 준비 url을 /live에서 /prepare로 변경
