+++ src/app/(post)/store/[storeId]/layout.tsx
@@ -1,13 +1,13 @@
 import { ReactNode } from 'react';
 
-import Header from '@/components/Header';
+import MainHeader from '@/components/Header';
 
 export default function StoreLayout({
   children,
 }: Readonly<{ children: ReactNode }>) {
   return (
     <>
-      <Header />
+      <MainHeader />
       {children}
     </>
   );
