--- src/app/live/prepare/page.module.scss
@@ -1,5 +1,0 @@
-@import '@/styles/variables/colors.scss';
-
-.container {
-  background-color: $bgColor;
-}
