+++ src/app/vote/page.module.scss
